Thermistor code is calibrated and working
The RawADC value was being measured relative to 5v, while we were
assuming it was measured relative to 3.3v. To calibrate the
temperatures, Ben found the datasheet for the thermistor, and it
turns out to be an NTCLP100E3103H thermistor. Using that, and
switching from the Steinhart-Hart equation to the Beta equation,
the temperature readings are now correct, +/- 2 degrees F or so.